+++ include/color.h
@@ -1,8 +1,5 @@
-/*
- * This is free and unencumbered software released into the public domain.
- *
- * For more information, please refer to <https://unlicense.org>
- */
+#ifndef COLOR_H
+# define COLOR_H
 
 //Regular text
 #define BLK "\e[0;30m"
@@ -75,6 +72,7 @@
 #define BHWHT "\e[1;97m"
 
 //Reset
-#define reset "\e[0m"
 #define CRESET "\e[0m"
 #define COLOR_RESET "\e[0m"
+
+#endif
