+++ include/minishell_src/minishell_init.h
@@ -1,0 +1,7 @@
+#ifndef MINISHELL_INIT_H
+# define MINISHELL_INIT_H
+
+void	init_minishell(void);
+void 	init_minishell_info(void);
+
+#endif
