+++ src/utils/chore_utils/is_white_space.c
@@ -1,4 +1,4 @@
-#include "../../include/minishell.h"
+#include "../../../include/minishell.h"
 
 int	is_white_space(char c)
 {
